fix: update network
